fetch/metadata/generated/window-history.sub.html

<!DOCTYPE html>
<!--
This test was procedurally generated. Please do not modify it directly.
Sources:
- fetch/metadata/tools/fetch-metadata.conf.yml
- fetch/metadata/tools/templates/window-history.sub.html
-->
<html lang="en">
  <meta charset="utf-8">
  <title></title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <body>
  <script>
  'use strict';

  function induceRequest(url, test, clear) {
    const win = window.open(url);

    test.add_cleanup(() => win.close());

    const whenDone = () => {
      return new Promise((resolve) => {
        addEventListener('message', (event) => {
          if (event.source === win) {
            resolve();
          }
        });
      })
    };

    return whenDone()
      .then(clear)
      .then(() => {
        win.history.back();
        return new Promise((resolve) => {
          win.addEventListener('unload', resolve);
        });
      })
      .then(whenDone);
  }

  const messageOpenerUrl = new URL(
    '/fetch/metadata/resources/message-opener.html', location
  );
  // For these tests to function, replacement must *not* be enabled during
  // navigation. Assignment must therefore take place after the document has
  // completely loaded [1]. This event is not directly observable, but it is
  // scheduled as a task immediately following the global object's `load`
  // event [2]. By queuing a task during the dispatch of the `load` event,
  // navigation can be consistently triggered without replacement.
  //
  // [1] https://html.spec.whatwg.org/multipage/history.html#location-object-setter-navigate
  // [2] https://html.spec.whatwg.org/multipage/parsing.html#the-end
  const responseParams = {
    mime: 'text/html',
    body: `<script>
      window.addEventListener('load', () => {
        setTimeout(() => location.assign('${messageOpenerUrl}'))
      });
    <`+`/script>`
  };

  promise_test((t) => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, ['sameOrigin', 'httpOrigin'], responseParams);

    return induceRequest(url, t, () => retrieve(key))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_not_own_property(headers, 'sec-fetch-site');
        });
  }, 'sec-fetch-site - HTTPS downgrade (header not sent) - ');

  promise_test((t) => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, ['httpOrigin', 'sameOrigin'], responseParams);

    return induceRequest(url, t, () => retrieve(key))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - HTTPS upgrade - ');

  promise_test((t) => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(key, ['sameOrigin', 'httpOrigin', 'sameOrigin'], responseParams);

    return induceRequest(url, t, () => retrieve(key))
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'sec-fetch-site - HTTPS downgrade-upgrade - ');
  </script>
  </body>
</html>
